separated logic for models
